Extracted to helper with the tests for the helper
First I extracted the ruby code contained within the ruby_block into a helper method. Normally I would place that helper method within a Module to namespace it with my company and include that within the recipe but that is for another commit.

The test is now using RSpec without ChefSpec to validate that when executing the ruby meethod it will set the global environment variable for the PATH.